Move GMT_GRID_HEADER and parser functions to pygmt/datatypes/header.py
